tests/requirements.txt
pytest==8.2.0
pytest-asyncio==0.23.7
pyyaml==6.0.1
httpx==0.27.0
litellm==1.40.0
fastapi==0.111.0
pydantic==2.7.0
testcontainers==4.14.2
docker==7.1.0
